+++ packages/traceloop-sdk/.prettierignore
@@ -1,2 +1,3 @@
 /dist
 /coverage
+/src/lib/generated
